chore: Rename Room Image To Chat Room
