+++ feature-main/room-create/src/main/java/com/seugi/roomcreate/screen/SecondScreen.kt
@@ -28,8 +28,8 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.unit.dp
 import com.seugi.designsystem.R
-import com.seugi.designsystem.component.RoomImageType
-import com.seugi.designsystem.component.SeugiRoomImage
+import com.seugi.designsystem.component.chat.ChatRoomType
+import com.seugi.designsystem.component.chat.SeugiChatRoom
 import com.seugi.designsystem.component.SeugiTopBar
 import com.seugi.designsystem.component.textfield.SeugiTextField
 import com.seugi.designsystem.theme.Black
@@ -84,8 +84,8 @@ fun SecondScreen(placeholder: String, onNameSuccess: (String) -> Unit, popBackSt
                 ),
         ) {
             Box(modifier = Modifier.align(Alignment.CenterHorizontally)) {
-                SeugiRoomImage(
-                    type = RoomImageType.ExtraLarge,
+                SeugiChatRoom(
+                    type = ChatRoomType.ExtraLarge,
                     text = placeholder.substring(0, 1),
                 )
                 Image(
